@@ -12,33 +12,32 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "7.0.0",
+    "@angular/animations": "8.0.2",
     "@angular/cdk": "^7.0.0",
-    "@angular/common": "7.0.0",
-    "@angular/compiler": "7.0.0",
-    "@angular/core": "7.0.0",
+    "@angular/common": "8.0.2",
+    "@angular/compiler": "8.0.2",
+    "@angular/core": "8.0.2",
     "@angular/flex-layout": "^7.0.0-beta.19",
-    "@angular/forms": "7.0.0",
-    "@angular/http": "7.0.0",
+    "@angular/forms": "8.0.2",
     "@angular/material": "^7.0.0",
-    "@angular/platform-browser": "7.0.0",
-    "@angular/platform-browser-dynamic": "7.0.0",
-    "@angular/router": "7.0.0",
+    "@angular/platform-browser": "8.0.2",
+    "@angular/platform-browser-dynamic": "8.0.2",
+    "@angular/router": "8.0.2",
     "chart.js": "^2.7.2",
     "core-js": "^2.5.6",
     "hammerjs": "^2.0.8",
-    "rxjs": "6.3.3",
-    "zone.js": "^0.8.26"
+    "rxjs": "6.5.2",
+    "zone.js": "~0.9.1"
   },
   "devDependencies": {
-    "@angular/cli": "^7.0.2",
-    "@angular/compiler-cli": "7.0.0",
-    "@angular-devkit/build-angular": "~0.10.0",
-    "@angular/language-service": "7.0.0",
+    "@angular/cli": "^8.0.4",
+    "@angular/compiler-cli": "8.0.2",
+    "@angular-devkit/build-angular": "~0.800.0",
+    "@angular/language-service": "8.0.2",
     "@types/jasmine": "^2.8.7",
     "@types/jasminewd2": "~2.0.3",
     "@types/node": "~10.12.0",
-    "codelyzer": "~4.5.0",
+    "codelyzer": "^5.0.1",
     "jasmine-core": "~3.2.1",
     "jasmine-spec-reporter": "~4.2.1",
     "karma": "~3.0.0",
@@ -49,6 +48,6 @@
     "protractor": "~5.4.1",
     "ts-node": "~7.0.1",
     "tslint": "~5.11.0",
-    "typescript": "3.1.3"
+    "typescript": "3.4.5"
   }
 }
